Small tweaks
- Separated database connection and "links" table handling in two files, functions stays the same.
- Returning errors from listenAndServe() instead of panicking directly.
- Separated the periodic generation of random tokens and the periodic garbage collection from main.go into a separate package. functions stays the same.
