+++ .golangci.yaml
@@ -36,6 +36,7 @@ linters-settings:
           - github.com/redds-be/reddlinks/internal/http
           - github.com/redds-be/reddlinks/internal/json
           - github.com/redds-be/reddlinks/internal/utils
+          - github.com/redds-be/reddlinks/internal/cron
           - github.com/lib/pq
           - github.com/mattn/go-sqlite3
           - github.com/google/uuid
