+++ main.go
@@ -20,8 +20,8 @@ import (
 	"database/sql"
 	"html/template"
 	"log"
-	"time"
 
+	"github.com/redds-be/reddlinks/internal/cron"
 	"github.com/redds-be/reddlinks/internal/database"
 	"github.com/redds-be/reddlinks/internal/env"
 	"github.com/redds-be/reddlinks/internal/http"
@@ -34,14 +34,28 @@ func main() {
 	envFile := ".env"
 	envVars := env.GetEnv(envFile)
 
-	dataBase, err := database.DBConnect(envVars.DBType, envVars.DBURL)
+	dbase, err := database.DBConnect(envVars.DBType, envVars.DBURL)
 	if err != nil {
 		log.Fatal(err)
 	}
 
+	// Defer the closing of the database connection
+	defer func(db *sql.DB) {
+		err := db.Close()
+		if err != nil {
+			log.Fatal(err)
+		}
+	}(dbase)
+
+	// Create the links table if it doesn't exist
+	err = database.CreateLinksTable(dbase, envVars.DefaultMaxLength)
+	if err != nil {
+		log.Panic(err)
+	}
+
 	// Create a struct to connect to the database and send the instance name and url to the handlers
 	conf := &utils.Configuration{
-		DB:                     dataBase,
+		DB:                     dbase,
 		PortSTR:                envVars.PortStr,
 		InstanceName:           envVars.InstanceName,
 		InstanceURL:            envVars.InstanceURL,
@@ -52,42 +66,18 @@ func main() {
 		Version:                "noVersion",
 	}
 
-	// Generate a new token every x time
-	go func(duration time.Duration) {
-		for {
-			http.Token = utils.RandomToken()
-			time.Sleep(duration)
-		}
-	}(3 * time.Hour) //nolint:gomnd
+	cron.StartJobs(*conf, envVars)
 
-	// Defer the closing of the database connection
-	defer func(db *sql.DB) {
-		err := db.Close()
-		if err != nil {
-			log.Fatal(err)
-		}
-	}(conf.DB)
-
-	// Create the links table if it doesn't exist
-	err = database.CreateLinksTable(conf.DB, envVars.DefaultMaxLength)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	// Periodically clean the database
-	go func(duration time.Duration) {
-		for {
-			err := conf.CollectGarbage()
-			if err != nil {
-				log.Println("Could not collect garbage:", err)
-			}
-			time.Sleep(duration)
-		}
-	}(time.Duration(envVars.TimeBetweenCleanups) * time.Minute)
-
+	// Parse html templates
 	http.Templates = template.Must(template.ParseFiles("static/index.html", "static/add.html",
 		"static/error.html", "static/pass.html", "static/privacy.html"))
 
+	// Create an adapter for the server
 	httpAdapter := http.NewAdapter(*conf)
-	httpAdapter.Run()
+
+	// Start the server
+	err = httpAdapter.Run()
+	if err != nil {
+		log.Panic(err)
+	}
 }
